Fix review issues with yac
 - Add handling of missing metadata

 - Move protobuf converters to separate file
