+++ irohad/consensus/yac/transport/impl/network_impl.cpp
@@ -15,55 +15,16 @@
  * limitations under the License.
  */
 
-#include "consensus/yac/impl/network_impl.hpp"
+#include "consensus/yac/transport/impl/network_impl.hpp"
 
 #include <grpc++/grpc++.h>
 
-#include "common/byteutils.hpp"
+#include "consensus/yac/transport/yac_pb_converters.hpp"
 #include "logger/logger.hpp"
 
 namespace iroha {
   namespace consensus {
     namespace yac {
-      proto::Vote serializeVote(const VoteMessage &vote) {
-        proto::Vote pb_vote;
-
-        auto hash = pb_vote.mutable_hash();
-        hash->set_block(vote.hash.block_hash);
-        hash->set_proposal(vote.hash.proposal_hash);
-
-        auto block_signature = hash->mutable_block_signature();
-        block_signature->set_signature(
-            vote.hash.block_signature.signature.to_string());
-        block_signature->set_pubkey(
-            vote.hash.block_signature.pubkey.to_string());
-
-        auto signature = pb_vote.mutable_signature();
-        signature->set_signature(vote.signature.signature.to_string());
-        signature->set_pubkey(vote.signature.pubkey.to_string());
-
-        return pb_vote;
-      }
-
-      nonstd::optional<VoteMessage> deserializeVote(
-          const proto::Vote &pb_vote) {
-        VoteMessage vote;
-        vote.hash.proposal_hash = pb_vote.hash().proposal();
-        vote.hash.block_hash = pb_vote.hash().block();
-        vote.hash.block_signature.signature =
-            *stringToBlob<iroha::sig_t::size()>(
-                pb_vote.hash().block_signature().signature());
-        vote.hash.block_signature.pubkey =
-            *stringToBlob<iroha::pubkey_t::size()>(
-                pb_vote.hash().block_signature().pubkey());
-        vote.signature.signature = *stringToBlob<iroha::sig_t::size()>(
-            pb_vote.signature().signature());
-        vote.signature.pubkey = *stringToBlob<iroha::pubkey_t::size()>(
-            pb_vote.signature().pubkey());
-
-        return vote;
-      }
-
       // ----------| Public API |----------
 
       NetworkImpl::NetworkImpl(const std::string &address,
@@ -147,6 +108,7 @@ namespace iroha {
         if (it == context->client_metadata().end()) {
           log_->error("Missing source address");
           // TODO handle missing source address
+          return grpc::Status::CANCELLED;
         }
         auto address = std::string(it->second.data(), it->second.size());
         auto peer = peers_addresses_.at(address);
@@ -168,6 +130,7 @@ namespace iroha {
         if (it == context->client_metadata().end()) {
           log_->error("Missing source address");
           // TODO handle missing source address
+          return grpc::Status::CANCELLED;
         }
         auto address = std::string(it->second.data(), it->second.size());
         auto peer = peers_addresses_.at(address);
@@ -194,6 +157,7 @@ namespace iroha {
         if (it == context->client_metadata().end()) {
           log_->error("Missing source address");
           // TODO handle missing source address
+          return grpc::Status::CANCELLED;
         }
         auto address = std::string(it->second.data(), it->second.size());
         auto peer = peers_addresses_.at(address);
